End of Day Push 5/30
Currently working on transitioning from the movement phase to the action phase. On the Unit script, there is an extra class that is called ActionStateController, which is where all of the Action stuff will go. It's currently a monobehavior, but I'm not sure if I can delete that soon.

Also need to remember that after the Action state is done, I'll need to transition to the End phase, which is more or less a buffer phase that I think I can keep through the enemy turn. We'll see, this is new territory and I'm not fully sure of my capabilities.